CO batch functionality develop

+++ src/main/java/com/jrtp/ies/dc/model/CaseDetail.java
@@ -9,5 +9,5 @@ import lombok.EqualsAndHashCode;
 @EqualsAndHashCode(callSuper=false)
 public class CaseDetail extends Applicant {
 
-	private short caseId;
+	private Integer caseId;
 }
